Allow setting hub_connect_port explicitly
And rename hub_ip_connect to hub_connect_ip, to be more
consistent. When running inside kubernetes, it is possible
that we port map a Service object, so that the port you have
to hit to access it isn't the same port the hub / proxy is
listening on. Having the hub_connect_port be explicit makes
this easy to specify