Use reflector pattern
We need to keep state within JupyterHub about the current running
status of all pods. We were currently doing this by making one
API request per pod every `poll_interval` - this is quite expensive,
and makes jupyterhub's polling scale linearly with number of
total users on the hub. Ideally we want this to be as close to
constant time as possible.

With this commit, we'll instead start an extra thread that'll
perform a 'list and watch' operation on the types of resources
we are interested in, and make sure that it is kept up to date
as atomically as possible (although there might be some lag sometime).
This makes poll be a simple synchronous dictionary access, rather
than a much more complex network poll.

This should make everything much faster - polling was one of the
biggest consumers of network time without this patch, and that
should significantly reduce with this patch.

We also take this opportunity to encode better the conditions
in which the pod is no longer considered running - particularly
when it is in 'Terminating' stage. This was a little bit flaky
before, but is ok now!