Removed hack for security in k8s <= v1.5
Pre k8s 1.6, the `automountServiceAccountToken` option was not available in pod specification, so a hack was used to disable auto mounting of a serviceAccountToken. Such token could be used to access the kubernetes master api-server as far as I understand. I think the 'disable-api-access' hack worked by making a dummy mount to the same path either overriding the previously auto-mounted path or prohibiting a future mount on the path.