removed `heritage` label selection
Label selection on all labels may end up a bit much, so in the zero-to-jupyterhub-k8s repo I reduced it to `component`, `app`, and `release`. Kubespawner is not helm dependent and should not be linked to `release` though, and having custom configuration of label selection could be troublesome, as all singleuser-servers could require a restart or get their labels updated if the kubespawner setting was updated. In the future, perhaps we will tackle this?