Require kubernetes client >= 8 (k8s 1.12 support)
Using minikube to test k8s 1.11 is a bit troublesome and caused test
failures, but k8s 1.11 is quite old now that 1.16 is out. I dropped
testing support of 1.11 and started require kubespawner to be installed
alongside a kubernetes client that has k8s 1.12 support. This does not
mean that we are forced to work on k8s 1.12 clusters as a client is
typically backward compatible.

It is very similar to saying, we need to require a newer version of
kubectl, but you don't have to upgrade your kubernetes cluster.

NOTE: Issues about minikube together with k8s 1.11 relates to socat
and cri-o isn't installed etc, it is not trivial to fix but it can be
fixxed if we really want it.