Fix downstream issue of serialization
This fixes what was reported in z2jh, where the hub doesn't crash but
fails to render the progress properly when it is serializing the events
and passing them to the browser to render them.

The error we got then was:

```
    TypeError: Object of type 'V1Event' is not JSON serializable
```

I think it makes best sense to not require jupyterhub to know about the
structure of the content and downstream make this serializable, but
instead reaquire all things to be serializable.

ref: https://github.com/jupyterhub/zero-to-jupyterhub-k8s/issues/1501