Refactor: let make_pod's names match kubespawner's config
KubeSpawner expose configuration with certain names that sadly doesn't
align so well with k8s. This commit makes sure that the make_pod helper
function doesn't introduce a third variation on how to name the same
thing.

Practically:

- KubeSpawner `gid` -> make_pod `run_as_gid` -> K8s `runAsGroup`.
- KubeSpawner `uid` -> make_pod `run_as_uid` -> K8s `runAsUser`.
- KubeSpawner `privileged` -> make_pod `run_privileged` -> K8s
`privileged`

This commit makes the middle convention align with the KubeSpawner
convention, so `run_as_gid` and become `gid` for example.