Terminate process correctly from reflector thread
Our reflectors run on their own thread, and sys.exit() from
a thread does not actually exit the process - just the thread.
When the reflector has to shut down because it can not talk
to the k8s master anymore for a while, we had just been stopping
the reflector thread and doing nothing else.

This will instead actually cause the hub to stop, and in z2jh
restart - which seems the right thing to do.

Ref https://github.com/2i2c-org/pilot-hubs/issues/680,
where we encountered this during a period of time when the
k8s master was unreachable for about 1 minute.