Allow a dropdown of images users can choose from in profile_list
The most common customization that admins want is the ability to
separate the image the user launches from from their resources
(CPU, RAM, GPU, etc). Instead of offering a full matrix of options
as suggested in https://github.com/jupyterhub/kubespawner/issues/307,
this PR *just* allows configurable dropdown of images for the user
to choose from.

I also clean-up the visual appearance of the profile list a little,
aligning things here and there :)

Ref https://github.com/jupyterhub/kubespawner/issues/307