Generalize 'options' to allow overriding any traitlet
Introduces a new 'options' parameter to profile, that allows
rendering a set of <select> tags in the profile_list page for
any option the administrator might want to allow the user to
choose from. This allows further customization of the spawned
server without leading to a combinatorial explosion of
profiles.

Each 'choice' in the options has its own set of kubespawner_overrides,
as different choices often necessitate setting multiple kubespawner
traitlet properties - for example, image choices might want to
affect default_url, while memory / CPU limits might affect machine
type selection. The simplest way to deal with this was to introduce
this slightly nested and repetitive config structure, as it is
simple to reason about.

This might mean many hub admins might get by with just a single
profile with a bunch of 'options'